grouped commands / facades together
