+++ packages/hedera-wallet-snap/packages/snap/src/commands/account/CreateAccountCommand.ts
@@ -23,7 +23,7 @@ import {
   SimpleTransfer,
   TxReceipt,
   TxReceiptExchangeRate,
-} from '../types/hedera';
+} from '../../types/hedera';
 import {
   AccountCreateTransaction,
   AccountId,
@@ -33,7 +33,7 @@ import {
   TransactionReceipt,
   TransferTransaction,
 } from '@hashgraph/sdk';
-import { CryptoUtils } from '../utils/CryptoUtils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
 
 export class CreateAccountCommand {
   readonly #currentBalance: AccountBalance;
