+++ packages/hedera-wallet-snap/packages/snap/src/commands/account/FreezeAccountCommand.ts
