+++ packages/hedera-wallet-snap/packages/snap/src/commands/allowance/ApproveAllowanceCommand.ts
@@ -18,17 +18,17 @@
  *
  */
 
-import { ApproveAllowanceAssetDetail } from '../types/params';
-import { TxReceipt } from '../types/hedera';
+import { ApproveAllowanceAssetDetail } from '../../types/params';
+import { TxReceipt } from '../../types/hedera';
 import {
   AccountAllowanceApproveTransaction,
   AccountId,
   Client,
   Hbar,
 } from '@hashgraph/sdk';
-import { Utils } from '../utils/Utils';
-import { CryptoUtils } from '../utils/CryptoUtils';
-import { EMPTY_STRING } from '../types/constants';
+import { Utils } from '../../utils/Utils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
+import { EMPTY_STRING } from '../../types/constants';
 
 export class ApproveAllowanceCommand {
   readonly #spenderAccountId: string;
