+++ packages/hedera-wallet-snap/packages/snap/src/facades/account/DeleteAccountFacade.ts
@@ -18,17 +18,17 @@
  *
  */
 
-import { SnapDialogParams, WalletSnapParams } from '../types/state';
-import { DeleteAccountRequestParams } from '../types/params';
-import { TxReceipt, AccountBalance } from '../types/hedera';
+import { SnapDialogParams, WalletSnapParams } from '../../types/state';
+import { DeleteAccountRequestParams } from '../../types/params';
+import { TxReceipt, AccountBalance } from '../../types/hedera';
 import { heading, text } from '@metamask/snaps-ui';
-import { SnapUtils } from '../utils/SnapUtils';
+import { SnapUtils } from '../../utils/SnapUtils';
 import { providerErrors } from '@metamask/rpc-errors';
 
-import { AccountInfo, Account } from '../types/account';
-import { SnapState } from '../snap/SnapState';
-import { HederaClientImplFactory } from '../client/HederaClientImplFactory';
-import { DeleteAccountCommand } from '../commands/DeleteAccountCommand';
+import { AccountInfo, Account } from '../../types/account';
+import { SnapState } from '../../snap/SnapState';
+import { HederaClientImplFactory } from '../../client/HederaClientImplFactory';
+import { DeleteAccountCommand } from '../../commands/account/DeleteAccountCommand';
 
 export class DeleteAccountFacade {
   /**
