+++ packages/hedera-wallet-snap/packages/snap/src/facades/allowance/ApproveAllowanceFacade.ts
@@ -21,16 +21,16 @@
 import {
   ApproveAllowanceAssetDetail,
   ApproveAllowanceRequestParams,
-} from '../types/params';
-import { MirrorTokenInfo, TxReceipt } from '../types/hedera';
+} from '../../types/params';
+import { MirrorTokenInfo, TxReceipt } from '../../types/hedera';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
 import { providerErrors } from '@metamask/rpc-errors';
-import { SnapDialogParams, WalletSnapParams } from '../types/state';
-import { SnapUtils } from '../utils/SnapUtils';
-import { CryptoUtils } from '../utils/CryptoUtils';
-import { HederaClientImplFactory } from '../client/HederaClientImplFactory';
-import { ApproveAllowanceCommand } from '../commands/ApproveAllowanceCommand';
+import { SnapDialogParams, WalletSnapParams } from '../../types/state';
+import { SnapUtils } from '../../utils/SnapUtils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
+import { HederaClientImplFactory } from '../../client/HederaClientImplFactory';
+import { ApproveAllowanceCommand } from '../../commands/allowance/ApproveAllowanceCommand';
 
 export class ApproveAllowanceFacade {
   public static async approveAllowance(
