+++ packages/hedera-wallet-snap/packages/snap/src/index.ts
@@ -24,11 +24,11 @@ import { OnRpcRequestHandler } from '@metamask/snaps-types';
 import { divider, heading, panel, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
 import { SignMessageCommand } from './commands/SignMessageCommand';
-import { ApproveAllowanceFacade } from './facades/ApproveAllowanceFacade';
-import { DeleteAccountFacade } from './facades/DeleteAccountFacade';
-import { DeleteAllowanceFacade } from './facades/DeleteAllowanceFacade';
-import { GetAccountBalanceFacade } from './facades/GetAccountBalanceFacade';
-import { GetAccountInfoFacade } from './facades/GetAccountInfoFacade';
+import { ApproveAllowanceFacade } from './facades/allowance/ApproveAllowanceFacade';
+import { DeleteAccountFacade } from './facades/account/DeleteAccountFacade';
+import { DeleteAllowanceFacade } from './facades/allowance/DeleteAllowanceFacade';
+import { GetAccountBalanceFacade } from './facades/account/GetAccountBalanceFacade';
+import { GetAccountInfoFacade } from './facades/account/GetAccountInfoFacade';
 import { StakeHbarFacade } from './facades/StakeHbarFacade';
 import { TransferCryptoFacade } from './facades/TransferCryptoFacade';
 import { AssociateTokensFacade } from './facades/hts/AssociateTokensFacade';
@@ -37,7 +37,7 @@ import { CreateTokenFacade } from './facades/hts/CreateTokenFacade';
 import { DeleteTokenFacade } from './facades/hts/DeleteTokenFacade';
 import { DissociateTokensFacade } from './facades/hts/DissociateTokensFacade';
 import { EnableKYCAccountFacade } from './facades/hts/EnableKYCAccountFacade';
-import { FreezeAccountFacade } from './facades/hts/FreezeAccountFacade';
+import { FreezeAccountFacade } from './facades/account/FreezeAccountFacade';
 import { MintTokenFacade } from './facades/hts/MintTokenFacade';
 import { PauseTokenFacade } from './facades/hts/PauseTokenFacade';
 import { WipeTokenFacade } from './facades/hts/WipeTokenFacade';
